feat: v3.0.2
特性：
1.实时图表支持查看核心使用率，CPU每个核心的频率，内存与SWAP占用细节（点击MEM图表查看详情）
2.图表历史数据增加为20s区间，数据更丰富
3.优化数据获取性能，资源占用更低，性能更佳

+++ app/src/main/java/com/minikano/f50_sms/utils/ShellKano.kt
@@ -476,17 +476,18 @@ class ShellKano {
             }
         }
 
-        fun executeShellFromAssetsSubfolder(context: Context, assetSubPath: String): String? {
+        fun executeShellFromAssetsSubfolder(context: Context, assetSubPath: String,outFileName:String="tmp_script.sh"): String? {
             try {
                 val assetManager = context.assets
 
-                // 从 assets/shell/ 复制，比如 assetSubPath = "shell/test.sh"
                 val inputStream = assetManager.open(assetSubPath)
-                val outFile = File(context.filesDir, "temp_script.sh")
+                val outFile = File(context.filesDir, outFileName)
 
-                inputStream.use { input ->
-                    FileOutputStream(outFile).use { output ->
-                        input.copyTo(output)
+                if (!outFile.exists()) {
+                    inputStream.use { input ->
+                        FileOutputStream(outFile).use { output ->
+                            input.copyTo(output)
+                        }
                     }
                 }
